requirements.txt
fastapi==0.68.0
uvicorn==0.15.0
python-dotenv==0.19.0
sqlalchemy==1.4.23
psycopg2-binary==2.9.1
alembic==1.7.1
python-jose[cryptography]==3.3.0
passlib==1.7.4
bcrypt==3.2.0
python-multipart==0.0.5
email-validator==1.1.3
pypdf2==3.0.1
python-magic==0.4.27
beautifulsoup4==4.12.2
openai==0.28.1
pinecone-client==3.0.0
numpy==1.24.3